requirements.txt
torch>=1.13.0
torchvision>=0.14.0
opencv-python>=4.7.0.72
pandas>=1.5.3
openpyxl>=3.1.2
matplotlib>=3.7.1
Pillow>=9.4.0
ultralytics>=8.0.0